(PC-23190)[API] chore: extract get event test
