+++ api/tests/routes/public/individual_offers/v1/endpoints_test.py
@@ -16,162 +16,6 @@ from . import utils
 
 
 @pytest.mark.usefixtures("db_session")
-class GetEventTest:
-    def test_404_when_requesting_a_product(self, client):
-        api_key = offerers_factories.ApiKeyFactory()
-        event_offer = offers_factories.ThingOfferFactory(venue__managingOfferer=api_key.offerer)
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-
-        assert response.status_code == 404
-        assert response.json == {"event_id": ["The event offer could not be found"]}
-
-    def test_get_event(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        product = offers_factories.ProductFactory(thumbCount=1)
-        event_offer = offers_factories.EventOfferFactory(
-            subcategoryId=subcategories.SEANCE_CINE.id,
-            venue=venue,
-            description="Un livre de contrepèterie",
-            name="Vieux motard que jamais",
-            product=product,
-        )
-        event_offer_id = event_offer.id
-
-        num_query = 1  # feature flag WIP_ENABLE_OFFER_CREATION_API_V1
-        num_query += 1  # retrieve API key
-        num_query += 1  # retrieve offer
-
-        with testing.assert_num_queries(num_query):
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/events/{event_offer_id}"
-            )
-
-        assert response.status_code == 200
-        assert response.json == {
-            "accessibility": {
-                "audioDisabilityCompliant": False,
-                "mentalDisabilityCompliant": False,
-                "motorDisabilityCompliant": False,
-                "visualDisabilityCompliant": False,
-            },
-            "bookingContact": None,
-            "bookingEmail": None,
-            "categoryRelatedFields": {"author": None, "category": "SEANCE_CINE", "stageDirector": None, "visa": None},
-            "description": "Un livre de contrepèterie",
-            "enableDoubleBookings": False,
-            "externalTicketOfficeUrl": None,
-            "eventDuration": None,
-            "id": event_offer.id,
-            "image": {
-                "credit": None,
-                "url": f"http://localhost/storage/thumbs/products/{human_ids.humanize(product.id)}",
-            },
-            "itemCollectionDetails": None,
-            "location": {"type": "physical", "venueId": event_offer.venueId},
-            "name": "Vieux motard que jamais",
-            "status": "SOLD_OUT",
-            "ticketCollection": None,
-            "priceCategories": [],
-        }
-
-    def test_event_with_not_selectable_category_can_be_retrieved(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(
-            venue=venue,
-            subcategoryId=subcategories.DECOUVERTE_METIERS.id,
-        )
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-
-        assert response.status_code == 200
-        assert response.json["categoryRelatedFields"] == {"category": "DECOUVERTE_METIERS", "speaker": None}
-
-    def test_get_show_offer_without_show_type(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(
-            subcategoryId=subcategories.SPECTACLE_REPRESENTATION.id,
-            venue=venue,
-        )
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-        assert response.status_code == 200
-        assert response.json["categoryRelatedFields"] == {
-            "author": None,
-            "category": "SPECTACLE_REPRESENTATION",
-            "performer": None,
-            "showType": None,
-            "stageDirector": None,
-        }
-
-    def test_get_music_offer_without_music_type(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(
-            subcategoryId=subcategories.CONCERT.id,
-            venue=venue,
-        )
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-        assert response.status_code == 200
-        assert response.json["categoryRelatedFields"] == {
-            "author": None,
-            "category": "CONCERT",
-            "performer": None,
-            "musicType": None,
-        }
-
-    def test_ticket_collection_by_email(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(
-            venue=venue,
-            withdrawalType=offers_models.WithdrawalTypeEnum.BY_EMAIL,
-            withdrawalDelay=259201,  # 3 days + 1 second
-        )
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-
-        assert response.status_code == 200
-        assert response.json["ticketCollection"] == {"daysBeforeEvent": 3, "way": "by_email"}
-
-    def test_ticket_collection_on_site(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(
-            venue=venue,
-            withdrawalType=offers_models.WithdrawalTypeEnum.ON_SITE,
-            withdrawalDelay=1801,  # 30 minutes + 1 second
-        )
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-
-        assert response.status_code == 200
-        assert response.json["ticketCollection"] == {"minutesBeforeEvent": 30, "way": "on_site"}
-
-    def test_ticket_collection_no_ticket(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(
-            venue=venue,
-            withdrawalType=offers_models.WithdrawalTypeEnum.NO_TICKET,
-        )
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/events/{event_offer.id}"
-        )
-
-        assert response.status_code == 200
-        assert response.json["ticketCollection"] is None
-@pytest.mark.usefixtures("db_session")
 class GetEventsTest:
     ENDPOINT_URL = "http://localhost/public/offers/v1/events"
